+++ src/components/QRCodeScanner.jsx
