+++ src/css/BarIndicator.css
