+++ src/css/QrReader.css
